requirements.txt
biopython>=1.83
parasail>=1.3.4
pandas>=2.2
matplotlib>=3.8
